requirements.txt
# Web Scraping & Crawling
# crawl4ai==0.8.6
# playwright==1.49.0

# AI Intelligence
# groq==0.5.0
# python-dotenv==1.0.1

# Reporting & Email
# reportlab==4.1.0
# yagmail==0.15.293

httpx==0.27.2
beautifulsoup4==4.12.3
groq==0.5.0
python-dotenv==1.0.1
reportlab==4.1.0
yagmail==0.15.293
